add class Role fixes #79, fixes #78, fixes #77

+++ segfault-api/src/main/java/heigvd/ch/segfaultapi/repositories/RoleRepoitory.java
@@ -2,5 +2,5 @@ package heigvd.ch.segfaultapi.repositories;
 
 import heigvd.ch.segfaultapi.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface RoleRepoitory extends JpaRepository<Role, Long>{
+public interface RoleRepoitory extends JpaRepository<Role, Integer>{
 }
